test(noise)+fix: 9-chunk golden gate + Java-shift UB fix
게이트를 golden 9청크 전부로 확장 (55741 라인 0 diff) — c.1.-1 이 글로벌
픽커 용암(y<-54) 경로를, c.0.-1 이 추가 aquifer 물을 커버한다. 뮤테이션
프로브 결과: RNG 순서/해시 키/코너 매핑/표면 +8 은 게이트가 잡고 (111~
4599 diffs), flat_cache 창 폴백·OF/WS 술어 구분·lerp3-점진 FP 순서·vein
RNG 체인은 이 골든에서 값-비가시 (문서화된 커버리지 경계, 구현은 바닐라
그대로 유지).

UBSan 이 잡은 음수 int 좌시프트 (Java 는 래핑 정의, C 는 UB) 를 무부호
우회로 수정: aquifer fromGrid/서프 오프셋, psl 쿼트 정렬, flat_cache
테이블 좌표. 값 변화 없음 (release 게이트 9청크 0 diff 유지).

+++ core/src/noise_chunk.c
@@ -58,9 +58,10 @@ static int32_t mth_floor(double d) {
 /* --- preliminarySurfaceLevel (Long2IntMap 메모 — 값-중립) --- */
 
 int32_t hc_nc_psl(hc_noise_chunk_t *nc, int32_t x, int32_t z) {
-    /* QuartPos.toBlock(QuartPos.fromBlock(v)) = (v >> 2) << 2 */
-    int32_t qx = (x >> 2) << 2;
-    int32_t qz = (z >> 2) << 2;
+    /* QuartPos.toBlock(QuartPos.fromBlock(v)) = (v >> 2) << 2 — Java 의
+     * 음수 좌시프트는 래핑이라 무부호로 우회한다 (값 범위상 동일) */
+    int32_t qx = (int32_t)((uint32_t)(x >> 2) << 2);
+    int32_t qz = (int32_t)((uint32_t)(z >> 2) << 2);
     /* ColumnPos.asLong(x, z) = (x & 0xFFFFFFFF) | (z << 32) */
     uint64_t key = (uint64_t)(uint32_t)qx | ((uint64_t)(uint32_t)qz << 32);
     uint32_t h = (uint32_t)(key ^ (key >> 29)) * 2654435761u;
@@ -317,9 +318,9 @@ int hc_nc_init(hc_noise_chunk_t *nc, hc_arena_t *arena,
         hc_df_flat_t *fl = &nc->flat[f];
         int32_t       child = g->nodes[fl->node].a;
         for (int32_t qi = 0; qi <= nc->noise_size_xz; qi++) {
-            int32_t bx = (nc->first_noise_x + qi) << 2; /* QuartPos.toBlock */
+            int32_t bx = (int32_t)((uint32_t)(nc->first_noise_x + qi) << 2);
             for (int32_t qj = 0; qj <= nc->noise_size_xz; qj++) {
-                int32_t bz = (nc->first_noise_z + qj) << 2;
+                int32_t bz = (int32_t)((uint32_t)(nc->first_noise_z + qj) << 2);
                 fl->values[qi + qj * flat_size] = nc_eval(
                     nc, child, (double)bx, 0.0, (double)bz, HC_DF_MODE_SP);
             }
